+++ src/main/java/janggi/moveStrategy/JumpStrategy.java
@@ -1,11 +1,11 @@
 package janggi.moveStrategy;
 
 import janggi.board.Board;
+import janggi.moveStrategy.rule.MoveVector;
+import janggi.moveStrategy.rule.MovingRule;
+import janggi.moveStrategy.rule.MovingRules;
 import janggi.piece.Team;
 import janggi.position.Position;
-import janggi.rule.MoveVector;
-import janggi.rule.MovingRule;
-import janggi.rule.MovingRules;
 
 public class JumpStrategy implements MoveStrategy {
