+++ src/test/java/janggi/position/PositionTest.java
@@ -3,7 +3,7 @@ package janggi.position;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import janggi.rule.MoveVector;
+import janggi.moveStrategy.rule.MoveVector;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
